% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediafire_gen.R
\name{mediafire_gen}
\alias{mediafire_gen}
\title{Mediafire Link Generator}
\usage{
mediafire_gen(link)
}
\arguments{
\item{link}{Shared Mediafire link}
}
\value{
Generated Mediafire link for download
}
\description{
Mediafire Link Generator
}
\examples{
mediafire_gen("http://www.imdb.com/title/tt1490017/")
}
